add: ログイン状態に応じてページ遷移する処理を useNavigateUser としてhookに分けた #62
